Revert "added component folder learn-react-navigation-20th-April"
This reverts commit 3001a91c6f3ce0c48d2d0b7f3edd1a042df41bf8.
